# Conflicts:
#	marketing-check/src/main/resources/pre-istio/scheduler.xml
#	marketing-check/src/main/resources/pre/scheduler.xml
#	marketing-check/src/main/resources/prod-istio/scheduler.xml
#	marketing-check/src/main/resources/prod/scheduler.xml
#	marketing-check/src/main/resources/scheduler.xml
#	marketing-check/src/main/resources/sim/scheduler.xml
#	marketing-cs/src/main/java/com/br/marketing/speedconfig/MarketingCommonConfig.java

+++ marketing-cs/src/main/java/com/br/marketing/service/Impl/PushDataServiceImpl.java
@@ -1228,7 +1228,7 @@ public class PushDataServiceImpl implements PushDataService {
             int xieChengCount = 1;
             if (isJson(data)) {
                 JSONObject jsonObject = JSONObject.parseObject(data);
-                id = Long.valueOf(jsonObject.getInteger("localId"));
+                id = jsonObject.getLong("localId");
             } else {
                 id = Long.valueOf(data);
                 localFile = localFileMapper.selectByPrimaryKey(id);
@@ -1274,7 +1274,7 @@ public class PushDataServiceImpl implements PushDataService {
     public void  pushXieChengSmsCollidingToDbData(String data) {
         try {
             JSONObject jsonObject = JSONObject.parseObject(data);
-            Long localId = Long.valueOf(jsonObject.getInteger("localId"));
+            Long localId = jsonObject.getLong("localId");
             Boolean isNewFile = jsonObject.getBooleanValue("isNewFile");
             // 创建线程池
             ThreadPoolExecutor xieChengSmsCollidingThread =
